Use factored out mv library in classification
