+++ classification/test_cnn.py
@@ -6,9 +6,9 @@ from sklearn.metrics import classification_report
 from torch.utils.data import DataLoader
 from torchvision import models
 from tqdm import tqdm
-from utils.data_loader import Resisc45Loader
-from utils.utils import get_label_list
-from utils.utils import parse_config
+from myrtle_vision.utils.data_loader import Resisc45Loader
+from myrtle_vision.utils.utils import get_label_list
+from myrtle_vision.utils.utils import parse_config
 
 
 def get_cnn(num_classes, model_name, weights_path):
