+++ src/myrtle_vision/utils/data_loader.py
@@ -4,8 +4,8 @@ import random
 import torch.utils.data
 from PIL import Image
 from torchvision import transforms
-from utils.utils import get_label_number
-from utils.utils import load_imagepaths_and_labels
+from myrtle_vision.utils.utils import get_label_number
+from myrtle_vision.utils.utils import load_imagepaths_and_labels
 
 
 class Resisc45Loader(torch.utils.data.Dataset):
