+++ src/myrtle_vision/vit/distill.py
@@ -5,7 +5,7 @@ import torch.nn.functional as F
 from torch import nn
 from torch.quantization import DeQuantStub
 from torch.quantization import QuantStub
-from vit_pytorch.vit_pytorch import ViT
+from myrtle_vision.vit.vit_pytorch import ViT
 
 
 class DistillableViT(ViT):
